count unique cases modified
