+++ corehq/apps/callcenter/_design/views/case_actions_by_user/reduce.js
